add job details view
