+++ jobby/server/jobby/urls.py
@@ -2,8 +2,10 @@
 
 from django.urls import path
 from jobby.views.stateViews import get_all_states, populate_state_database
-from jobby.views.jobViews import get_all_jobs, job_detail
+from jobby.views.jobViews import job_view
 from jobby.views.scrapeViews import scrape_by_url
+from jobby.views.jobDetailViews import job_detail
+
 
 
 # specifies the url patterns
@@ -11,7 +13,7 @@ from jobby.views.scrapeViews import scrape_by_url
 urlpatterns = [
     path('states/', get_all_states, name='getAllStates'),
     path('states/create/', populate_state_database, name='createStates'),
-    path('jobs/', get_all_jobs, name='getAllJobs'),
+    path('jobs/', job_view, name='jobView'),
     path('jobs/<pk>/', job_detail, name = 'jobDetails'),
     path('scrape/', scrape_by_url, name="scrapeByUrl")
 ]
